project_2/docker/streamlit/requirements.txt
streamlit
streamlit-lottie
Pillow
requests